Integrated NewsAPI feature
Integrated a NewsAPI feature that finds the latest news, also updated the news page design and when you click on a news tile it takes you to the url source.

Still needs work: Make it so its cybersecurity related news only and put a refresh button to grab the latest news.